server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
numpy==1.26.4
pydantic==2.8.2
transformers==4.44.0
accelerate==0.34.0
ruff==0.6.9
